Fix formatting and add image for DataLad dandisets organization
Without empty line, list was rendered in a single paragraph.

I also extended text a bit to point to super dataset and some more useful pointers
on datalad/git.

Screenshot helps to orient users a bit more so they see something familiar
(on github) and that it is welcoming with its README.md